Glassmorphic AR hazard overlays, bbox support, README updates
- DOM-based HazardOverlay component with backdrop-filter blur, severity
  colors, corner brackets, and minimal label pills
- Gemini highlight_hazard tool now accepts box_2d coordinates
- Backend converts Gemini 0-1000 bbox to normalized 0-1 for frontend
- BBox type updated to float64 for normalized coordinates
- Overlay toggle: voice command in AR mode, O key in CCTV/smartphone
- Shared types across hooks and components (removed duplicate interfaces)
- Web Speech API type declarations for TypeScript strict mode
- README: How It Works section, 5 concrete use cases, AR mode description fix

+++ pkg/types/types.go
@@ -33,11 +33,12 @@ type DetectedObject struct {
 }
 
 // BBox defines a bounding box for visual overlays.
+// Coordinates are normalized 0–1 (fraction of frame dimensions).
 type BBox struct {
-	X      int `json:"x"`
-	Y      int `json:"y"`
-	Width  int `json:"width"`
-	Height int `json:"height"`
+	X      float64 `json:"x"`
+	Y      float64 `json:"y"`
+	Width  float64 `json:"width"`
+	Height float64 `json:"height"`
 }
 
 // Hazard represents a detected safety issue.
